feat(settings): aba Sistema dedicada
Move autostart, allowScripts (kill-switch perfil), auto-update toggle +
UpdateCard, scriptHistoryEnabled, export/import config e a seleção de
idioma da aba Aparência pra uma nova aba "Sistema". Aparência fica só
com tema, ThemeCustomizer e o botão "Definir como ativo". Sem fieldset
interno na nova aba — os blocos ficam direto na section, espaçados pelo
gap. Closes #35.
