+++ src/locales/en.json
@@ -106,6 +106,7 @@
       "tabs": "Tabs",
       "appearance": "Appearance",
       "shortcut": "Shortcut",
+      "system": "System",
       "history": "History"
     },
     "profile": {
