+++ src/locales/pt-BR.json
@@ -106,6 +106,7 @@
       "tabs": "Abas",
       "appearance": "Aparência",
       "shortcut": "Atalho",
+      "system": "Sistema",
       "history": "Histórico"
     },
     "profile": {
