+++ src/settings/SectionTabs.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { useTranslation } from "react-i18next";
 
-export type Section = "tabs" | "appearance" | "shortcut" | "history";
+export type Section = "tabs" | "appearance" | "shortcut" | "system" | "history";
 
-const SECTIONS: Section[] = ["tabs", "appearance", "shortcut", "history"];
+const SECTIONS: Section[] = ["tabs", "appearance", "shortcut", "system", "history"];
 
 export interface SectionTabsProps {
   active: Section;
