+++ src/settings/__tests__/AppearanceSection.test.tsx
@@ -5,27 +5,16 @@ import { I18nextProvider } from "react-i18next";
 import { createI18n } from "../../core/i18n";
 import { AppearanceSection } from "../AppearanceSection";
 import type { Theme } from "../../core/types/Theme";
-import type { Language } from "../../core/types/Language";
 
 async function renderSection(overrides: Partial<{
   theme: Theme;
-  language: Language;
-  autostart: boolean;
   onThemeChange: (t: Theme) => void;
-  onLanguageChange: (l: Language) => void;
-  onAutostartChange: (e: boolean) => void;
   onSetActiveProfile: () => void;
-  onExportConfig: () => void;
-  onImportConfig: () => void;
 }> = {}) {
   const i18n = await createI18n("pt-BR");
   const props = {
     theme: "dark" as Theme,
-    language: "auto" as Language,
-    autostart: false,
     onThemeChange: vi.fn(),
-    onLanguageChange: vi.fn(),
-    onAutostartChange: vi.fn(),
     ...overrides,
   };
   const utils = render(
@@ -52,19 +41,6 @@ describe("AppearanceSection", () => {
     expect(props.onThemeChange).toHaveBeenCalledWith("light");
   });
 
-  it("pre-selects the current language in the select", async () => {
-    await renderSection({ language: "en" });
-    const select = screen.getByLabelText(/idioma/i) as HTMLSelectElement;
-    expect(select.value).toBe("en");
-  });
-
-  it("calls onLanguageChange when a different language is selected", async () => {
-    const user = userEvent.setup();
-    const { props } = await renderSection({ language: "auto" });
-    await user.selectOptions(screen.getByLabelText(/idioma/i), "ptBr");
-    expect(props.onLanguageChange).toHaveBeenCalledWith("ptBr");
-  });
-
   it("does not render the set-active button when onSetActiveProfile is undefined", async () => {
     await renderSection();
     expect(screen.queryByTestId("set-active-profile")).toBeNull();
@@ -78,42 +54,5 @@ describe("AppearanceSection", () => {
     expect(btn).toBeTruthy();
     await user.click(btn);
     expect(onSetActiveProfile).toHaveBeenCalledTimes(1);
-  });
-
-  it("autostart checkbox reflects the current value", async () => {
-    await renderSection({ autostart: true });
-    const cb = screen.getByTestId("autostart-toggle") as HTMLInputElement;
-    expect(cb.checked).toBe(true);
-  });
-
-  it("calls onAutostartChange when the checkbox is toggled", async () => {
-    const user = userEvent.setup();
-    const { props } = await renderSection({ autostart: false });
-    await user.click(screen.getByTestId("autostart-toggle"));
-    expect(props.onAutostartChange).toHaveBeenCalledWith(true);
-  });
-
-  it("does not render export/import buttons when callbacks are absent", async () => {
-    await renderSection();
-    expect(screen.queryByTestId("export-config")).toBeNull();
-    expect(screen.queryByTestId("import-config")).toBeNull();
-  });
-
-  it("renders and invokes the export-config callback when clicked", async () => {
-    const user = userEvent.setup();
-    const onExportConfig = vi.fn();
-    await renderSection({ onExportConfig });
-    const btn = screen.getByTestId("export-config");
-    await user.click(btn);
-    expect(onExportConfig).toHaveBeenCalledTimes(1);
-  });
-
-  it("renders and invokes the import-config callback when clicked", async () => {
-    const user = userEvent.setup();
-    const onImportConfig = vi.fn();
-    await renderSection({ onImportConfig });
-    const btn = screen.getByTestId("import-config");
-    await user.click(btn);
-    expect(onImportConfig).toHaveBeenCalledTimes(1);
   });
 });
